Added auth so that only specific roles could trigger specific actions, in case of naughty ppl

+++ StatybuWeb/Controllers/AdminController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using StatybuWeb.Services.Api;
 using System.Security.Claims;
 using StatybuWeb.Dto;
+using StatybuWeb.Services.Auth0;
 
 namespace StatybuWeb.Controllers
 {
@@ -14,10 +15,24 @@ namespace StatybuWeb.Controllers
     {
         private readonly IAzureService _azureService;
         private readonly IAzureBlobStorageService _azureBlobStorageService;
-        public AdminController(IAzureService azureService, IAzureBlobStorageService azureBlobStorageService)
+        private readonly IAuth0Service _auth0Service;
+
+        public AdminController(IAzureService azureService, IAzureBlobStorageService azureBlobStorageService, IAuth0Service auth0Service)
         {
             _azureService = azureService;
             _azureBlobStorageService = azureBlobStorageService;
+            _auth0Service = auth0Service;
+        }
+
+        private bool IsUserAuthorized(string role)
+        {
+            string? userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return false;
+            }
+            var userRoles = _auth0Service.GetUserRoles(userId).Result.Select(x => x.Name);
+            return userRoles.Contains(role);
         }
 
         [Authorize(AuthenticationSchemes = "Auth0")]
@@ -26,48 +41,58 @@ namespace StatybuWeb.Controllers
             return View(await _azureBlobStorageService.GetImagesFilesFromBlobStorage());
         }
 
-        [Authorize(AuthenticationSchemes = "Auth0", Roles ="Admin")]
+        [Authorize(AuthenticationSchemes = "Auth0")]
         [HttpPost]
         public async Task<ActionResult> UploadImage(IFormFile file)
         {
-            if (file != null && file.Length > 0)
+            if (IsUserAuthorized("Admin"))
             {
-                await _azureBlobStorageService.UploadFileToBlobStorage(file);
-            }
-
-            return RedirectToAction("Index");
-        }
-
-        [Authorize(AuthenticationSchemes = "Auth0", Roles = "Admin")]
-        public async Task<ActionResult> ImageActions()
-        {
-            return View(await _azureBlobStorageService.GetImagesFilesFromBlobStorage());
-        }
-
-        [Authorize(AuthenticationSchemes = "Auth0", Roles = "Admin")]
-        [HttpPost]
-        public async Task<ActionResult> DeleteImages(List<Picture> fileNames)
-        {
-            BlobContainerClient containerClient = await _azureBlobStorageService.GetAzureBlobContainerClientFromSecrets();
-            foreach (var picture in fileNames)
-            {
-                if (picture.Selected)
+                if (file != null && file.Length > 0)
                 {
-                    try
-                    {
-                        // Get a reference to the file
-                        BlobClient blobClient = containerClient.GetBlobClient(picture.Name);
-
-                        // Delete the file
-                        await blobClient.DeleteAsync();
-                    }
-                    catch (Exception ex)
-                    {
-                        Console.WriteLine($"Error deleting file': {ex.Message}");
-                    }
+                    await _azureBlobStorageService.UploadFileToBlobStorage(file);
                 }
             }
             return RedirectToAction("Index");
+        }
+
+        [Authorize(AuthenticationSchemes = "Auth0")]
+        public async Task<ActionResult> ImageActions()
+        {
+            if (IsUserAuthorized("Admin"))
+            {
+                return View(await _azureBlobStorageService.GetImagesFilesFromBlobStorage());
+            }
+            return Unauthorized();
+        }
+
+        [Authorize(AuthenticationSchemes = "Auth0")]
+        [HttpPost]
+        public async Task<ActionResult> DeleteImages(List<Picture> fileNames)
+        {
+            if (IsUserAuthorized("Admin"))
+            {
+                BlobContainerClient containerClient = await _azureBlobStorageService.GetAzureBlobContainerClientFromSecrets();
+                foreach (var picture in fileNames)
+                {
+                    if (picture.Selected)
+                    {
+                        try
+                        {
+                            // Get a reference to the file
+                            BlobClient blobClient = containerClient.GetBlobClient(picture.Name);
+
+                            // Delete the file
+                            await blobClient.DeleteAsync();
+                        }
+                        catch (Exception ex)
+                        {
+                            Console.WriteLine($"Error deleting file': {ex.Message}");
+                        }
+                    }
+                }
+                return RedirectToAction("Index");
+            }
+            return Unauthorized();
         }
     }
 }
